@@ -75,6 +75,7 @@
     "@mui/material": "5.15.5",
     "@next/env": "14.1.0",
     "@parcel/watcher": "^2.4.0",
+    "@unts/patch-package": "^8.0.0",
     "concurrently": "8.2.2",
     "cross-env": "^7.0.3",
     "dotenv": "16.3.1",
@@ -82,12 +83,11 @@
     "graphql": "^16.8.1",
     "next": "14.1.0",
     "next-sitemap": "4.2.3",
-    "patch-package": "8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.2",
     "sharp": "0.33.2",
-    "webpack": "5.89.0"
+    "webpack": "5.93.0"
   },
   "devDependencies": {
     "@graphcommerce/eslint-config-pwa": "8.0.6",
