@@ -31,11 +31,10 @@
     "graphql": "^16.8.1",
     "next": "14.1.0",
     "next-sitemap": "4.2.3",
-    "patch-package": "8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sharp": "0.33.2",
-    "webpack": "5.89.0"
+    "webpack": "5.93.0"
   },
   "devDependencies": {
     "@graphcommerce/eslint-config-pwa": "8.0.6",
