@@ -29,7 +29,7 @@
     "next": "14.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "webpack": "5.89.0"
+    "webpack": "5.93.0"
   },
   "devDependencies": {
     "@graphcommerce/eslint-config-pwa": "8.0.6",
